Move units file to sub-package, yank unused code
- Remove `InList()` function since we are not
  actively using it

- Move `units` file to sub-package (for now)

refs #12
